Consolidate Field and data type subclasses. Add more Python stubs
