+++ cpp/src/arrow/table/column.cc
@@ -20,7 +20,7 @@
 #include <memory>
 #include <sstream>
 
-#include "arrow/field.h"
+#include "arrow/type.h"
 #include "arrow/util/status.h"
 
 namespace arrow {
