+++ cpp/src/arrow/table/schema-test.cc
@@ -20,7 +20,6 @@
 #include <string>
 #include <vector>
 
-#include "arrow/field.h"
 #include "arrow/table/schema.h"
 #include "arrow/type.h"
 #include "arrow/types/string.h"
