+++ cpp/src/arrow/table/schema.cc
@@ -22,7 +22,7 @@
 #include <sstream>
 #include <vector>
 
-#include "arrow/field.h"
+#include "arrow/type.h"
 
 namespace arrow {
