+++ cpp/src/arrow/type.cc
@@ -17,7 +17,16 @@
 
 #include "arrow/type.h"
 
+#include <sstream>
+#include <string>
+
 namespace arrow {
+
+std::string Field::ToString() const {
+  std::stringstream ss;
+  ss << this->name << " " << this->type->ToString();
+  return ss.str();
+}
 
 const std::shared_ptr<NullType> NA = std::make_shared<NullType>();
 const std::shared_ptr<BooleanType> BOOL = std::make_shared<BooleanType>();
@@ -31,5 +40,28 @@ const std::shared_ptr<Int32Type> INT32 = std::make_shared<Int32Type>();
 const std::shared_ptr<Int64Type> INT64 = std::make_shared<Int64Type>();
 const std::shared_ptr<FloatType> FLOAT = std::make_shared<FloatType>();
 const std::shared_ptr<DoubleType> DOUBLE = std::make_shared<DoubleType>();
+const std::shared_ptr<StringType> STRING = std::make_shared<StringType>();
+
+std::string ListType::ToString() const {
+  std::stringstream s;
+  s << "list<" << value_type->ToString() << ">";
+  if (!this->nullable) {
+    s << " not null";
+  }
+  return s.str();
+}
+
+std::string StructType::ToString() const {
+  std::stringstream s;
+  s << "struct<";
+  for (size_t i = 0; i < fields_.size(); ++i) {
+    if (i > 0) s << ", ";
+    const std::shared_ptr<Field>& field = fields_[i];
+    s << field->name << ": " << field->type->ToString();
+  }
+  s << ">";
+  if (!nullable) s << " not null";
+  return s.str();
+}
 
 } // namespace arrow
