+++ cpp/src/arrow/types/struct.cc
@@ -17,24 +17,6 @@
 
 #include "arrow/types/struct.h"
 
-#include <cstdlib>
-#include <memory>
-#include <sstream>
-#include <string>
-
 namespace arrow {
-
-std::string StructType::ToString() const {
-  std::stringstream s;
-  s << "struct<";
-  for (size_t i = 0; i < fields_.size(); ++i) {
-    if (i > 0) s << ", ";
-    const Field& field  = fields_[i];
-    s << field.name << ": " << field.type->ToString();
-  }
-  s << ">";
-  if (!nullable) s << " not null";
-  return s.str();
-}
 
 } // namespace arrow
